Design: 타임라인 UI 구현 완료

+++ domain/src/main/java/com/mangpo/domain/model/RecordEntity.kt
@@ -5,5 +5,5 @@ data class RecordEntity(
     val keyword: String,
     val content: String? = null,
     val date: String,
-    val star: Double
+    val star: Float
 )
